Cleanup of Transaction hashes
