+++ src/wrapped_radix_engine_toolkit/low_level/mod.rs
@@ -3,16 +3,16 @@ mod dummy_types;
 
 mod compiled_notarized_intent;
 mod execution_summary;
+mod intent_signature;
+mod intent_signatures;
 mod manifest_summary;
 mod message;
 mod notarized_transaction;
+mod notary_signature;
 mod public_key_hash;
 mod signed_intent;
 mod transaction_classes;
-// mod transaction_hash;
-mod intent_hash;
-mod notary_signature;
-mod signed_intent_hash;
+mod transaction_hashes;
 mod transaction_header;
 mod transaction_intent;
 mod transaction_manifest;
@@ -21,6 +21,8 @@ mod transaction_receipt;
 pub use compiled_notarized_intent::*;
 pub use dummy_types::*;
 pub use execution_summary::*;
+pub use intent_signature::*;
+pub use intent_signatures::*;
 pub use manifest_summary::*;
 pub use message::*;
 pub use notarized_transaction::*;
@@ -28,9 +30,7 @@ pub use notary_signature::*;
 pub use public_key_hash::*;
 pub use signed_intent::*;
 pub use transaction_classes::*;
-// pub use transaction_hash::*;
-pub use intent_hash::*;
-pub use signed_intent_hash::*;
+pub use transaction_hashes::*;
 pub use transaction_header::*;
 pub use transaction_intent::*;
 pub use transaction_manifest::*;
